Fix build

+++ src/main/java/ir/sahab/kafka/reader/ParquetFile.java
@@ -41,7 +41,7 @@ public class ParquetFile<T extends Message> implements AutoCloseable {
 
         ParquetWriterBuilder builder =
                 new ParquetWriterBuilder(filePath).withConf(properties.hadoopConf)
-                        .withRowGroupSize(properties.blockSize)
+                        .withRowGroupSize((long) properties.blockSize)
                         .withCompressionCodec(properties.compressionCodecName)
                         .withWriteMode(Mode.OVERWRITE)
                         .withPageSize(properties.pageSize);
